Create a seam to add live lookup for Folio
